services/retrieval_service/requirements.txt
fastapi==0.115.6
langchain==0.3.12
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-postgres==0.0.12
langchain-unstructured==0.1.6
httpx==0.27.2
openai==1.57.4
pgvector==0.2.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
unstructured-client==0.28.1
uvicorn==0.34.0